refactored sign in and sign up routes and controllers
