+++ package.json
@@ -7,9 +7,9 @@
     "format": "prettier --ignore-path ./.gitignore --write",
     "prepare": "husky",
     "lint-staged": "lint-staged -v",
-    "setup:client": "cd client && npm i",
-    "setup:server": "cd server && npm i",
-    "setup": "npm i && npm run setup:client && npm run setup:server",
+    "setup:client": "cd client && npm install",
+    "setup:server": "cd server && npm install",
+    "setup": "npm install && npm run setup:client && npm run setup:server",
     "docker:client": "cd client && docker-compose up -d",
     "docker:server": "cd server && docker-compose up -d",
     "docker": "npm run docker:client && npm run docker:server",
