+++ server/package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start:app": "NODE_ENV=development nodemon server/index.js",
-    "postinstall": "CONT=$(docker ps -aqf name=node-app) && [ $CONT ] && docker exec -w /backend-app $CONT npm i || true",
+    "postinstall": "CONT=$(docker ps -aqf name=node-app) && [ $CONT ] && docker exec -w /backend-app $CONT npm install || true",
     "start": "docker-compose up",
     "build": "docker-compose up --build",
     "prune": "docker-compose down -v --rmi local",
