--- server/server/Routes/sign.js
@@ -1,7 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const signingController = require("../Controllers/signIn");
-
-router.post("/", signingController.signIn);
-
-module.exports = router;
